fixed 13.05

+++ Brain/FPSMeter.cs
@@ -9,7 +9,7 @@
         {
             get
             {
-                return (int)(int)fps;
+                return (int)fps;
             }
         }
